Update README.md
This project is a collaboration work between myself and two others for my Mobile Applications Development module. The concept was to create a basic prototype of an app that would work across devices to help students get to Maynooth University. With one student working on the UI and another working on the bus and train element, the Ride Share element was my primary focus while I also completed the conditional rendering and other design elements too. The train and bus searches receive data from a JSON that is included in the code while the Ride Share searches receive data from an API hosted in another repository named cs385Project.